feat: Implement chart export utility and enhance chart components
- Added ChartExportUtils for exporting charts in PNG, JPG, and PDF formats.
- Introduced ImprovedAxisTick component for better axis tick rendering.
- Created EmptyChart component to handle empty states in charts.
- Developed CustomLegend for improved legend display and interaction.
- Implemented CustomReferenceLabel for displaying reference lines in charts.
- Added CustomBar and CustomDot components for custom rendering of bars and dots in charts.
- Enhanced tooltip functionality with CustomGraphTooltip for better data representation.
- Configured chart settings in chartConfig.js for consistent styling.
- Implemented hooks for managing chart data and ticks.
- Improved color utility functions for dynamic color resolution based on theme.
- Added data normalization utility for processing chart data.
- Included formatters for consistent Y-axis tick formatting.
