+++ src/platform/src/common/features/airQuality-charts/ChartContainer/config/chartConfig.js
