+++ src/platform/src/common/features/airQuality-charts/ChartContainer/hooks/useChartDimensions.js
