+++ src/platform/src/common/features/airQuality-charts/ChartContainer/utils/chartExportUtils.js
