+++ src/platform/src/common/features/airQuality-charts/MoreInsightsChart/components/Series/SeriesComponents.jsx
@@ -1,0 +1,1 @@
+export { Line as SeriesLine, Bar as SeriesBar } from 'recharts';
