+++ src/platform/src/common/features/airQuality-charts/MoreInsightsChart/config/chartConfig.js
@@ -1,0 +1,13 @@
+export const CHART_CONFIG = {
+  chartMargin: { top: 50, right: 20, bottom: 30 },
+  exportMargin: { top: 5, right: 5, left: 15, bottom: 30 },
+  exportStyles: {
+    fontSize: 12,
+    fontFamily: 'system-ui, -apple-system, sans-serif',
+  },
+  yAxisLabel: {
+    fontSize: 12,
+    fontWeight: '600',
+    position: { dy: -40, dx: 5 },
+  },
+};
