+++ src/platform/src/common/features/airQuality-charts/MoreInsightsChart/hooks/useChartData.js
@@ -1,0 +1,1 @@
+export { useNormalizedData as default } from '../utils/dataUtils';
